feat: 시큐리티 필터 추가. 관리자페이지(admin)에서 요청 들어올시 관리자 권한 필터링.

+++ src/main/java/kr/co/itid/cms/service/auth/model/MenuPermissionData.java
@@ -38,13 +38,14 @@ public class MenuPermissionData implements Serializable {
      * @param targetPermission 확인할 권한
      * @return 권한이 있을 경우 true
      */
-    public boolean hasPermission(int userIdx, int level, String targetPermission) {
+    public boolean hasPermission(long userIdx, int level, String targetPermission) {
         boolean levelAllowed = false;
 
         for (Map.Entry<Integer, List<PermissionEntry>> entry : sortedPermissionMap.entrySet()) {
             for (PermissionEntry permissionEntry : entry.getValue()) {
                 // 1. ID 일치 시 즉시 true
-                if (permissionEntry.getUserIdx() != null && permissionEntry.getUserIdx().equals(userIdx)) {
+                if (permissionEntry.getUserIdx() != null &&
+                        permissionEntry.getUserIdx().equals(userIdx)) {
                     if (permissionEntry.getPermissions().contains(targetPermission)) {
                         return true;
                     }
@@ -63,7 +64,7 @@ public class MenuPermissionData implements Serializable {
     /**
      * 정렬 순서에 따라 권한 엔트리를 추가
      */
-    public void addPermissionEntry(int sort, Integer userIdx, Integer level, Set<String> permissions) {
+    public void addPermissionEntry(int sort, long userIdx, Integer level, Set<String> permissions) {
         sortedPermissionMap.computeIfAbsent(sort, k -> new ArrayList<>())
                 .add(new PermissionEntry(userIdx, level, permissions));
     }
